Config preprocessor: Add basic support for including YAML files
The support is quite basic, it really is only text processing and does not
care about the yaml syntax at all.

example use-case:
You can move the per-board-type information to a separate file that can
even be located in separate directory.

pluma-target.yml:
!include "pluma-imx8mm-evk.yml"
console:
  serial:
    port: /dev/ttyUSB0

pluma-imx8mm-evk.yml:
system:
  prompt_regex: "\\S+@\\S+:"
  credentials:
    login: root
    password: root

variables:
  blk_test_dev: "/dev/mmcblk4"

note: when the preprocessor looks for a file, it first tries in the directory
of the parent yaml file, then it looks for it in the include search paths.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@traphandler.com>